Remove PipeWire 0.3.x module
This isn't necessary in the first place as PipeWire 0.3.x is included in the 20.08 FreeDesktop runtime.
In the past we needed to add PipeWire 0.2.x as Electron<13 defaulted to only have support for PipeWire 0.2.x for screen sharing on Wayland.
According to Slack's logs (accessible via DEB/RPM) Slack should be on Electro 13/Chromium 91.