Update metadata, and rename to metainfo
Change component type from 'desktop' to 'desktop-application' desktop
is the older type identifier for desktop-applications and should not
be used for new metainfo files, unless compatibility with very old
AppStream tools (pre 2016) is still wanted.

Set the "project_group" attribute to be the same as "developer_name".

Change the ID to be the same as the application ID.

Remove period from the summary, to be more inline with formatting
guidelines.

Describe how to launch this software with a launchable attribute
that links to our desktop file.

Change update_contact attribute to reflect current maintainers.
Does not appear to support multiple entries so I've just added
myself.

Declare that Slack supports and has been tested with touch devices,
mouse and keyboard.

Set internet as a required field, and use the download speed for
5+ participants in a video call as the requirement for download
speed.

Update the releases to match https://slack.com/release-notes/linux
and bring back release notes. I'm aware of the guidelines Flathub
has on no "bug fixes and performance improvements", but it is to
be noted this is not our app, it's proprietary so we can't come up
with accurate release notes.

Move all attributes above the releases section so we don't have
to contiously scroll to view anything else but releases and notes.

Update the open age rating service to include money-purchasing,
Slack pushes heavily for you to get a paid plan. It will prevent
you from reading messages that are too old if you don't at least
have a Pro plan.