Use the snap package for updates
Slack didn't update the release notes for Linux, but did update
the snap. Switch to using the snap package to get updates so we
get them as soon as they come out.

Also looking to reintroduce the beta channel, and switch this
branch to use 'stable' at some point. Probably on the next beta
release.